pyproject.toml
[tool.poetry]
name = "homestead solar"
version = "0.1.0"
description = ""
authors = ["Zeke Zumbro <jezumbro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
loguru = "^0.6.0"
python-dotenv = "^1.0.0"
httpx = "^0.23.3"
pymongo = "^4.3.3"
pydantic-mongo = "^1.0.1"
requests = "^2.28.2"
pycryptodome = "^3.17"
more-itertools = "^9.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-watch = "^4.2.0"
isort = "^5.12.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
